pyproject.toml
[project]
name = "ai-sonification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi>=0.115.12",
    "gradio>=5.29.0",
    "langchain>=0.3.24",
    "langchain-openai>=0.3.14",
    "nest-asyncio>=1.6.0",
    "openai>=1.76.0",
    "pydantic-ai>=0.1.8",
    "pydantic-ai-slim[openai]>=0.1.8",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "uvicorn[standard]>=0.34.2",
]
